Move to latest version with public AssociationMapView member in CLUE

+++ DataFormats/L1ScoutingSoA/interface/IndexSoA.h
@@ -5,7 +5,7 @@
 
 namespace l1sc {
 
-  GENERATE_SOA_LAYOUT(IndexLayout, SOA_COLUMN(uint32_t, indexes))
+  GENERATE_SOA_LAYOUT(IndexLayout, SOA_COLUMN(uint16_t, indexes))
 
   using IndexSoA = IndexLayout<>;
